QUAD-116: Bug fixes as mentioned in PR Story/quad 116 sprint 5 #174
