+++ Quadriga/src/main/webapp/WEB-INF/views/settings/publicwebsitesettings.jsp
@@ -1,15 +1,6 @@
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
 <%@ taglib prefix="sec" uri="http://www.springframework.org/security/tags"%>
 <%@ taglib uri="http://www.springframework.org/tags" prefix="spring"%>
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+
 <title>Public Page Settings</title>
-</head>
-<body>
-
 <a href="${pageContext.servletContext.contextPath}/auth/workbench/projects/${project.projectId}/settings/editabout">Change About Project details</a>
-
-</body>
-</html>
