+++ Quadriga/src/main/webapp/WEB-INF/views/settings/saveAbout.jsp
@@ -1,12 +1,6 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Insert title here</title>
-</head>
-<body>
-Your information is successfully saved.
-</body>
-</html>
+
+<title>About Project - Saved</title>
+
+<article>Your information is successfully saved.</article>
